==> urban_mobility_analysis/01_Accidents_merges.ipynb <==
import pandas as pd
import numpy as np
import os
from tqdm import tqdm

accident = pd.read_excel('data/accident_total_2016-2020.xlsx', sheet_name = '편집_1')

conzone = pd.read_csv('data/info_total.csv', encoding = 'cp949')

start_end_point = conzone.iloc[:, [5, 6, 7, 8, 9, 10]].drop_duplicates()
start_end_point.head(2)

total_df = pd.merge(accident, start_end_point, how = 'left', left_on = ['노선명', '방향'], right_on = ['도로단축명', '종점명']).dropna(subset = ['도로명', '도로단축명', '도로표출명', '기점종점방향구분코드'])

conzone.head(2)

total_df.head(3)

total_df['도로명']

save_total_df_dir = 'data/accident_merged.csv'

# 도로단축명, 기종점방향구분코드가 conzone과 total_df가 같을 경우,
# total_df의 '이정'이 conzone의 시작노드_도로이정과 종료노드_도로이정 사이에 있으면 해당 콘존을 입력
# 기종점방향구분코드 : E이면 시작노드 < 종료노드, S이면 시작노드 > 종료노드임.

total_df['conzoneID'] = None #새로운 비어있는 컬럼 : 콘존ID
total_df['conzoneName'] = None # 콘존명

for i in tqdm(range(len(total_df))): # total_df의 i번째 행에 대하여
    
    # 먼저, total_df[i]와 같은 도로단축명, 기종점 방향구분코드를 가진 conzone만 걸러낸다.
    i_roadname = total_df['도로명'].iloc[i]
    i_direction = total_df['기점종점방향구분코드'].iloc[i]

    czdf = conzone[(conzone['도로명'] == i_roadname) & (conzone['기점종점방향구분코드'] == i_direction)]
    
    # i_direction이 E냐 S냐에 따라서 다른 if 절을 적용
    # conzone의 시작노드_도로이정, 종료노드_도로이정 사이에 '도로이정'이 있을 시 해당 콘존을 투입
    location = total_df['이정'].iloc[i]
    
    for j in range(len(czdf)):
        start_location = czdf['시작노드_도로이정'].iloc[j]
        end_location = czdf['종료노드_도로이정'].iloc[j]
        
        if i_direction == 'E':
            if location >= start_location and location < end_location:
                conzoneID = czdf['콘존ID'].iloc[j]
                conzoneName = czdf['콘존명'].iloc[j]
                
            else:
                pass
                
        elif i_direction == 'S':
            if location <= start_location and location > end_location:
                conzoneID = czdf['콘존ID'].iloc[j]
                conzoneName = czdf['콘존명'].iloc[j]
                
            else:
                pass
            
        else:
            pass
    
    total_df['conzoneID'].iloc[i] = conzoneID
    total_df['conzoneName'].iloc[i] = conzoneName

total_df.to_csv(save_total_df_dir, encoding = 'cp949')
print('PROCESSING FINISHED')
