==> urban_mobility_analysis/05_accident+violation.ipynb <==
# ACCIDENT + VIOLATION + DANGER INDEX 데이터셋 가공
* 사고 데이터
* 고속도로 위험지수 데이터
* 위반 데이터

import pandas as pd
import numpy as np
import os
from tqdm import tqdm

from math import radians, cos, sin, asin, sqrt

# 데이터 가공

## DANGER INDEX
* 길이에 따른 위험지수의 가중평균으로 Section(Conzone)별 위험지수를 산정하기
* Vehicle Type별로 pivot_table을 사용하여 분리하기

DangerIndex = pd.read_csv('data/Freeway_Danger_Index.csv', encoding = 'cp949')

DangerIndex.head(3)

DangerIndex = pd.pivot_table(DangerIndex.iloc[:, 2:],
                             index = ['start_x', 'start_y', 'end_x', 'end_y', 'Freeway', 'Section'],
                             columns = ['vehicle_type']).reset_index()

DangerIndex.columns = ['start_x', 'start_y', 'end_x', 'end_y', 'Freeway', 'Section',
              'grade_1', 'grade_2', 'grade_3', 'grade_4', 'value_1', 'value_2', 'value_3', 'value_4']

def haversine(lon1, lat1, lon2, lat2):
    """
    Calculate the great circle distance in kilometers between two points 
    on the earth (specified in decimal degrees)
    """
    # convert decimal degrees to radians 
    lon1, lat1, lon2, lat2 = map(radians, [lon1, lat1, lon2, lat2])

    # haversine formula 
    dlon = lon2 - lon1 
    dlat = lat2 - lat1 
    a = sin(dlat/2)**2 + cos(lat1) * cos(lat2) * sin(dlon/2)**2
    c = 2 * asin(sqrt(a)) 
    r = 6371 # Radius of earth in kilometers. Use 3956 for miles. Determines return value units.
    
    return c * r

DangerIndex['length'] = DangerIndex.apply(lambda x: haversine(x['start_x'], x['start_y'], x['end_x'], x['end_y']), axis = 1)

# 모든 차종에서 value, grade가 같으므로 하나로 변경함
DangerIndex = DangerIndex.drop(['grade_2', 'grade_3', 'grade_4', 'value_2', 'value_3', 'value_4'], axis = 'columns')

DangerIndex['valueLen'] = DangerIndex['value_1'] * DangerIndex['length']
DangerIndex.head(3)

pv = pd.pivot_table(DangerIndex,
                    index = ['Freeway', 'Section'],
                    values = ['length', 'valueLen'],
                    aggfunc = 'sum').reset_index()

pv['avg_idx'] = pv['valueLen'] / pv['length']
pv = pv.drop(['length', 'valueLen'], axis = 'columns')

pv.head(3)

pv.to_csv('data/Freeway_Danger_Index_edited.csv', encoding = 'cp949')

## VIOLATION
* 시간대(0~3, 3~6)별로 구분하여 합쳐서 지수를 재생성하기

violation = pd.read_csv('data/violations_merged_02.csv', encoding = 'cp949')
violation.head(3)

def timefunc(x):
    
    t = int(x[0:2])
    
    if  t >= 0 and t < 3:
        return '00-03'
    elif t >= 3 and t < 6:
        return '03-06'
    elif t >= 6 and t < 9:
        return '06-09'
    elif t >= 9 and t < 12:
        return '09-12'
    elif t >= 12 and t < 15:
        return '12-15'
    elif t >= 15 and t < 18:
        return '15-18'
    elif t >= 18 and t < 21:
        return '18-21'
    elif t >= 21 and t < 24:
        return '21-24'
    else:
        pass

violation['time'] = violation['oTime'].apply(timefunc)
violation.head(3)

violation.to_csv('data/violations_merged_02.csv', encoding = 'cp949')

## ACCIDENT

accident = pd.read_csv('data/accident_merged_20170730-.csv', encoding = 'cp949')
accident.head(3)

# 시간단위_3시간 << 변수를 00-03, 12-15, ... 단위로 일괄 변경하여 저장하기
def timefunc2(x):
    if x == '0~3':
        return '00-03'
    elif x == '3~6':
        return '03-06'
    elif x == '6~9':
        return '06-09'
    elif x == '9~12':
        return '09-12'
    elif x == '12~15':
        return '12-15'
    elif x == '15~18':
        return '15-18'
    elif x == '18~21':
        return '18-21'
    elif x == '21~24':
        return '21-24'
    else:
        pass

accident['시간단위_3시간'] = accident['시간단위_3시간'].apply(timefunc2)

accident.head(3)

accident.to_csv('data/accident_merged_20170730-_time-edited.csv', encoding = 'cp949')
