==> urban_mobility_analysis/06_DangerIndex.ipynb <==
# DangerIndex : 위험지수에 방향(direction) 적용하기

import pandas as pd
import numpy as np
import os
from tqdm import tqdm

sectionlist = pd.read_csv('data/DangerIndex/Freeway_Section_list_TAAS.csv', encoding = 'cp949')
sectionlist.head(3)

conzonelist = pd.read_csv('data/Basic/info_total.csv', encoding = 'cp949')
conzonelist.head(3)

czsec = pd.merge(sectionlist, conzonelist, how = 'left',
                left_on = ['도로단축명', '콘존명'], right_on = ['도로단축명', '콘존명'])

czsec = czsec.dropna(axis = 0)[['Freeway_TAAS', 'Section_TAAS', '도로단축명', '콘존명', '기점종점방향구분코드']].drop_duplicates()
czsec.head(3)

czsec.to_csv('data/DangerIndex/Freeway_Section_list_TAAS_direction.csv', encoding = 'cp949')

DangerIndex = pd.read_csv('data/DangerIndex/Freeway_Danger_Index_edited.csv', encoding = 'cp949')

DangerIndex.head(3)

DangerIndexMerged = pd.merge(DangerIndex, czsec, how = 'left',
                            left_on = ['Freeway', 'Section'],
                            right_on = ['Freeway_TAAS', 'Section_TAAS'])

DangerIndexMerged.head(3)

DangerIndexMerged.to_csv('data/DangerIndex/Freeway_Danger_Index_edited_merged.csv', encoding = 'cp949')
