==> urban_mobility_analysis/07_violation.ipynb <==
# VIOLATION : 시간대별 평균값을 양산할 필요

import pandas as pd
import numpy as np
import os

violation_dir = 'data/violations'
violation_file = 'violations_merged_02.csv'
violation_path = os.path.join(violation_dir, violation_file)

violation = pd.read_csv(violation_path, encoding = 'cp949')

violation.head(3)

pv = pd.pivot_table(
    violation,
    index = ['date', 'routeNo', 'direction', 'conzoneId', 'conzoneNm', 'busLaneGubun', 'time'],
    aggfunc = 'sum'
)

pv = pv.reset_index().drop(['Unnamed: 0', 'Unnamed: 0.1', 'Unnamed: 0.1.1', 'laneNum', 'total_ViolationRate', 'HV_ViolationRate'], axis = 'columns')

pv.head(3)

pv['total_ViolationRate'] = pv['violationTF']/pv['totalTF']
pv['HV_ViolationRate'] = pv['violationTF_HV']/pv['totalTF_HV']

pv = pv.fillna(0)

save_file = 'violations_merged_03.csv'
save_path = os.path.join(violation_dir, save_file)

pv.to_csv(save_path, encoding = 'cp949')
