==> urban_mobility_analysis/08_JOIN.ipynb <==
# JOIN : ACCIDENT + VIOLATION + DANGERINDEX
* 세 데이터를 융합해서 하나의 데이터프레임으로 만들기

import pandas as pd
import numpy as np
import os
from tqdm import tqdm

accident_dir = 'data/accident'
DangerIndex_dir = 'data/DangerIndex'
violation_dir = 'data/violations'

accident_file = 'accident_merged_20170730-_time-edited.csv'
DangerIndex_file = 'Freeway_Danger_Index_edited_merged.csv'
violation_file = 'violations_merged_03.csv'

accident_path = os.path.join(accident_dir, accident_file)
DangerIndex_path = os.path.join(DangerIndex_dir, DangerIndex_file)
violation_path = os.path.join(violation_dir, violation_file)

accident = pd.read_csv(accident_path, encoding = 'cp949')
DangerIndex = pd.read_csv(DangerIndex_path, encoding = 'cp949')
violation = pd.read_csv(violation_path, encoding = 'cp949')

# JOIN

## accident + Violation

accident.head(3)

violation.head(3)

acc_vio = pd.merge(
    accident, violation, how = 'left',
    left_on = ['사고일자.1', '시간단위_3시간', 'conzoneID', '기점종점방향구분코드', 'conzoneName'],
    right_on = ['date', 'time', 'conzoneId', 'direction', 'conzoneNm']
).dropna(subset = ['HV_ViolationRate'])

acc_vio

acc_vio.to_csv('data/FINAL_accident+violence.csv', encoding = 'cp949')

## DANGER + VIOLATION

violation.head(3)

DangerIndex.head(3)

dan_vio = pd.merge(
    violation, DangerIndex, how = 'left',
    left_on = ['conzoneNm', 'direction'],
    right_on = ['콘존명', '기점종점방향구분코드']
).dropna(subset = ['avg_idx'])

dan_vio.columns

dan_vio = dan_vio.drop(['Unnamed: 0_x', 'Unnamed: 0_y', 'Unnamed: 0.1', 'Freeway', 'Section', '콘존명', '기점종점방향구분코드'], axis = 'columns')

dan_vio.to_csv('data/FINAL_danger+violation.csv', encoding = 'cp949')
