==> urban_mobility_analysis/09_correlation.ipynb <==
# Correlation
* 고속도로 위험지수와 위반율 사이의 상관관계를 분석함

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt

import scipy.stats as stats
from statsmodels.formula.api import ols
from statsmodels.stats.anova import anova_lm

dan_vio = pd.read_csv('data/FINAL_danger+violation.csv', encoding = 'cp949')

# 데이터 타입 바꾸고 쓸모없는 변수 지우기
dan_vio = dan_vio.drop(['Unnamed: 0'], axis = 'columns')
dan_vio['date'] = dan_vio['date'].astype(str)
dan_vio['routeNo'] = dan_vio['routeNo'].astype(str)

dan_vio.head(3)

dan_vio.describe().round(2)

corr = dan_vio[['total_ViolationRate', 'HV_ViolationRate', 'avg_idx']].corr(method = 'pearson')
# 중차량의 위반비율과 고속도로 위험지수는 뚜렷한 양적 상관관계를 가진다!!
corr
